Change routing system

+++ src/cmps/chat/chat-dropdown.jsx
@@ -3,25 +3,29 @@
 import { useNavigate } from "react-router-dom"
 import { roomService } from "../../services/room.service"
 
-function ChatDropdown({ roomId, chatType, setIsOptionsSelected, setIsChatPopupShown, setIsGroupPopupShown }) {
+function ChatDropdown({ roomId, chatType, setIsOptionsSelected }) {
     const navigate = useNavigate()
 
-    // Change place to upload img 
-    // CHANGE GROUP NAME
-    // CHANGE PARTICIPANTS IN GROUP
-    // DELETE GROUP
-    // CHANGE GROUP FROM LAST SEEN TO PARTICIPANTS
+
 
     function onDeleteRoom() {
+        if(!window.confirm(`Are you sure you want to delete ${chatType}?`)) return
+        setIsOptionsSelected(prevState=>!prevState)
         navigate('/rooms')
         roomService.deleteRoom(roomId)
     }
     
+    function onGroupClick(){
+        setIsOptionsSelected(prevState=>!prevState)
+        // navigate(`/info`)
+        navigate(`/rooms/${roomId}/info`)
+    }
+
     return (
         <div className="dropdown" onClick={() => setIsOptionsSelected(prevState => !prevState)}>
             {/* <button onClick="myFunction()" className="dropbtn">Dropdown</button> */}
             <div className="dropdown-content">
-                {chatType === 'group' && <a onClick={() => setIsChatPopupShown(prevState => !prevState)}>Group info</a>}
+                {chatType === 'group' && <a onClick={onGroupClick}>Group info</a>}
                 {chatType === 'chat' && <a onClick={onDeleteRoom}>Delete chat</a>}
             </div>
         </div>
